Added command list-candidate-files.
list-candidate-files list the files that will be backed up along with
a summary of the size of the files (i.e. not the size of the resulting
backup). This way it is possible to "debug" your selection expressions
without doing a backup.
duplicity is pretty rigid about how arguments should look, it assumes
that if a command has one argument that argument MUST be a backend URL.
So in order to use a one argument command that is a local dir the
argument had to be a bit "bolted on".